Rename gltr_once header
